move some files and packages
errorcheck now part of utils
wal in folder backup (will move package later)
backup.go in folder and package backup
backups.go renamed to basebackups.go and new folder and package backup
