+++ cmd/root.go
@@ -39,7 +39,6 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
-	ec "github.com/xxorde/pgglaskugel/errorcheck"
 	util "github.com/xxorde/pgglaskugel/util"
 
 	"github.com/kardianos/osext"
@@ -347,7 +346,7 @@ func initConfig() {
 
 	// Check if needed tools are available
 	err := testTools(baseBackupTools)
-	ec.Check(err)
+	util.Check(err)
 
 	// Check if the configured backend is supported
 	if err := storage.CheckBackend(viper.GetString("backup_to")); err != nil {
@@ -409,7 +408,7 @@ func validatePgData(pgData string) (err error) {
 func reloadConfiguration(db *sql.DB) (err error) {
 	query := "SELECT pg_reload_conf();"
 	_, err = db.Query(query)
-	ec.Check(err)
+	util.Check(err)
 	return err
 }
 
@@ -417,7 +416,7 @@ func reloadConfiguration(db *sql.DB) (err error) {
 func getPgSetting(db *sql.DB, setting string) (value string, err error) {
 	query := "SELECT setting FROM pg_settings WHERE name = $1;"
 	row := db.QueryRow(query, setting)
-	ec.Check(err)
+	util.Check(err)
 	err = row.Scan(&value)
 	if err != nil {
 		log.Fatal("Can not get PostgreSQL setting: ", setting, " err:", err)
@@ -584,7 +583,7 @@ func compressEncryptStream(input *io.ReadCloser, name string, storageBackend sto
 	// Watch output on stderror
 	compressDone := make(chan struct{}) // Channel to wait for WatchOutput
 	compressStderror, err := compressCmd.StderrPipe()
-	ec.Check(err)
+	util.Check(err)
 	go util.WatchOutput(compressStderror, log.Info, compressDone)
 
 	// Pipe the backup in the compression
@@ -624,7 +623,7 @@ func compressEncryptStream(input *io.ReadCloser, name string, storageBackend sto
 		gpgCmd.Stdin = compressStdout
 		// Watch output on stderror
 		gpgStderror, err := gpgCmd.StderrPipe()
-		ec.Check(err)
+		util.Check(err)
 		go util.WatchOutput(gpgStderror, log.Warn, gpgDone)
 
 		// Start encryption
