+++ cmd/setup.go
@@ -27,9 +27,9 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"github.com/xxorde/pgglaskugel/util"
 
 	log "github.com/Sirupsen/logrus"
-	ec "github.com/xxorde/pgglaskugel/errorcheck"
 	// This is needed but never directly called
 	_ "github.com/lib/pq"
 )
@@ -69,7 +69,7 @@ var (
 
 			// Check if needed tools are available
 			err := testTools(setupTools)
-			ec.Check(err)
+			util.Check(err)
 
 			// When no archive command set, set it
 			if viper.GetString("archive_command") == "" {
@@ -109,11 +109,11 @@ var (
 
 			// Get and check version via SQL
 			pgVersion, err := checkPgVersion(db)
-			ec.Check(err)
+			util.Check(err)
 
 			// Get version of the data
 			pgDataVersion, err := getMajorVersionFromPgData(pgData)
-			ec.Check(err)
+			util.Check(err)
 
 			log.WithFields(log.Fields{
 				"pgData":           pgData,
@@ -132,12 +132,12 @@ var (
 
 			// Create directories for backups, WAL and configuration
 			err = createDirs(archiveDir, subDirs)
-			ec.Check(err)
+			util.Check(err)
 
 			// Configure PostgreSQL for archiving
 			log.Info("Configure PostgreSQL for archiving.")
 			changed, _ := configurePostgreSQL(db, pgSettings)
-			ec.Check(err)
+			util.Check(err)
 			// If more than 0 setings have been changed we reload the configuration
 			if changed > 0 {
 				log.Info("Going to reload the configuration.")
@@ -145,7 +145,7 @@ var (
 
 				// Configure PostgreSQL again to see if all settings are good now!
 				changed, _ = configurePostgreSQL(db, pgSettings)
-				ec.Check(err)
+				util.Check(err)
 			}
 
 			if changed > 0 {
@@ -185,7 +185,7 @@ func init() {
 // it then shows the user the need to restart PostgreSQL
 func pgRestartDB(pgData string) (err error) {
 	postmasterPID, err := getPostmasterPID(pgData)
-	ec.Check(err)
+	util.Check(err)
 	log.Warn("Please restart PostgreSQL wth PID ", postmasterPID)
 	return err
 }
@@ -220,12 +220,12 @@ func configurePostgreSQL(db *sql.DB, settings map[string]string) (changed int, e
 	for setting := range settings {
 		settingShould := settings[setting]
 		settingIs, err := getPgSetting(db, setting)
-		ec.Check(err)
+		util.Check(err)
 		log.Debug(setting, " should be: ", settingShould, "   it is: ", settingIs)
 
 		if settingIs != settingShould {
 			err := setPgSetting(db, setting, settingShould)
-			ec.Check(err)
+			util.Check(err)
 			changed++
 		}
 	}
