+++ util/errorcheck.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package errorcheck
+package util
 
 import log "github.com/Sirupsen/logrus"
